Add new fab command line tool
This creates a new version of the fab command line tool which is
intended to mirror the way make works.  The design picks up a target
file and executes the FabRecipe class inside it.  This should make it
broadly compatible with other extensions, e.g. Baf.

The command includes improved and extended logging, but this has not
yet been extended to entirety of the fab library.  Eventually, the
command interface will also be extended to make use of the various
progress indicators available through the rich library.